Add comments and restructuration
