+++ README.md
@@ -1,2 +1,19 @@
 # google-2.0
 Implementation of a search engine from scratch
+
+We are working on two given collections:
+
+- CACM collection
+- CS276 collection
+
+## Task 1
+
+### Linguistic processing
+
+Entry point: `CACMIndex.py` and `CS276Index.py`.
+
+Helper functions:
+
+- `textProcessing.py`
+- `indexBuilder.py`
+- `CACMParser.py`
